# Remove the unnecessary laziness