A case class derivation has become faster.

current master

```
[info] Benchmark                                         (size)   Mode  Cnt    Score   Error   Units
[info] DecoderBench.decodeSelf                               10  thrpt   20  366.514 ± 6.585  ops/ms
[info] DecoderBench.decodeSelf                               30  thrpt   20  102.063 ± 1.192  ops/ms
[info] DecoderBench.decodeToCaseClass                        10  thrpt   20  298.655 ± 7.995  ops/ms
[info] DecoderBench.decodeToCaseClass                        30  thrpt   20   89.264 ± 1.342  ops/ms
[info] DecoderBench.decodeToCaseClassCachedDecoder           10  thrpt   20  355.724 ± 4.713  ops/ms
[info] DecoderBench.decodeToCaseClassCachedDecoder           30  thrpt   20  105.665 ± 1.411  ops/ms
[info] DecoderBench.decodeToMap                              10  thrpt   20  269.529 ± 2.802  ops/ms
[info] DecoderBench.decodeToMap                              30  thrpt   20   88.865 ± 1.237  ops/ms
[info] PutEncoderBench.encodeFromCaseClass                   10  thrpt   20  453.066 ± 10.942  ops/ms
[info] PutEncoderBench.encodeFromCaseClass                   30  thrpt   20  153.302 ±  3.588  ops/ms
[info] PutEncoderBench.encodeFromCaseClassCachedEncoder      10  thrpt   20  658.304 ± 16.896  ops/ms
[info] PutEncoderBench.encodeFromCaseClassCachedEncoder      30  thrpt   20  230.562 ±  6.825  ops/ms
[info] PutEncoderBench.encodeFromMap                         10  thrpt   20  696.779 ± 14.908  ops/ms
[info] PutEncoderBench.encodeFromMap                         30  thrpt   20  238.708 ±  5.202  ops/ms
```

this PR

```
[info] Benchmark                                         (size)   Mode  Cnt    Score    Error   Units
[info] DecoderBench.decodeSelf                               10  thrpt   20  354.381 ± 13.161  ops/ms
[info] DecoderBench.decodeSelf                               30  thrpt   20  103.059 ±  1.842  ops/ms
[info] DecoderBench.decodeToCaseClass                        10  thrpt   20  329.908 ±  4.829  ops/ms
[info] DecoderBench.decodeToCaseClass                        30  thrpt   20  101.625 ±  2.160  ops/ms
[info] DecoderBench.decodeToCaseClassCachedDecoder           10  thrpt   20  352.103 ±  8.697  ops/ms
[info] DecoderBench.decodeToCaseClassCachedDecoder           30  thrpt   20  105.546 ±  2.291  ops/ms
[info] DecoderBench.decodeToMap                              10  thrpt   20  232.200 ± 21.345  ops/ms
[info] DecoderBench.decodeToMap                              30  thrpt   20   82.773 ±  7.637  ops/ms
[info] PutEncoderBench.encodeFromCaseClass                   10  thrpt   20  585.058 ± 11.152  ops/ms
[info] PutEncoderBench.encodeFromCaseClass                   30  thrpt   20  212.122 ±  3.560  ops/ms
[info] PutEncoderBench.encodeFromCaseClassCachedEncoder      10  thrpt   20  671.182 ± 11.830  ops/ms
[info] PutEncoderBench.encodeFromCaseClassCachedEncoder      30  thrpt   20  234.327 ±  5.075  ops/ms
[info] PutEncoderBench.encodeFromMap                         10  thrpt   20  694.034 ± 21.357  ops/ms
[info] PutEncoderBench.encodeFromMap                         30  thrpt   20  241.655 ±  3.476  ops/ms
```